% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule_column.R
\name{applyrule.rule_column}
\alias{applyrule.rule_column}
\title{Prints the data frame in an html page and shows it.}
\usage{
\method{applyrule}{rule_column}(rule, finalformat, x, ...)
}
\arguments{
\item{rule}{A rule_column object}

\item{finalformat}{}

\item{x}{The data to which format has to be applied}
}
\value{
finalformat
}
\description{
Prints the data frame in an html page and shows it.
}
\examples{
data(iris)
library(dplyr)
condformat(iris) \%>\% rule_column(column="Species", css="background: red")
}

