% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html.R
\name{cf_field_to_css}
\alias{cf_field_to_css}
\title{How to export a cf_field to CSS}
\usage{
cf_field_to_css(cf_field, xview, css_fields, unlocked)
}
\arguments{
\item{cf_field}{A cf_field object. This is like a rule, but with the computed
colour values. It usually maps one-to-one to a CSS field.}

\item{xview}{A data frame with the columns to be printed and rows filtered}

\item{css_fields}{A list of matrices. The names of the list are CSS attributes and
each matrix is of the size of xview and contains the respective CSS values.}

\item{unlocked}{A logical matrix of cells unlocked (that can still be modified by further
rules).}
}
\value{
A list with two elements: css_fields and unlocked (with updated values)
}
\description{
This method is exported so package users can generate their own rules
}
