% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_latex.R
\name{cf_field_to_latex}
\alias{cf_field_to_latex}
\title{How to export cf values to latex}
\usage{
cf_field_to_latex(cf_field, xview, unlocked)
}
\arguments{
\item{cf_field}{A cf_field object. This is like a rule, but with the computed
colour values. It usually maps one-to-one to a CSS field.}

\item{xview}{A data frame with the columns to be printed and rows filtered}

\item{unlocked}{A logical matrix of cells unlocked (that can still be modified by further
rules).}
}
\value{
A list with two character matrices named \code{before} and \code{after}. Both
of these matrices must be of the same size as \code{xview}.
}
\description{
How to export cf values to latex
}
