% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condformat.R
\name{condformat}
\alias{condformat}
\title{Converts a data frame or matrix to a condformat_tbl class.
condformat_tbl objects allow to add conditional formatting information,
that can be viewed when the condformat_tbl object is printed.}
\usage{
condformat(x, select, col.names, ...)
}
\arguments{
\item{x}{A matrix or data.frame}

\item{select}{a character string or vector with the column names to be printed.
Columns not listed here may be used in conditional rules, although
will not appear in the generated table when printing. Default: colnames(x)}

\item{col.names}{a character string or vector of the same length of `select` with the desired column names to be shown
in the printed table.}

\item{...}{additional arguments passed to htmlTable}
}
\value{
The condformat_tbl object, that can be printed to generate the output.
}
\description{
They also have all properties from conventional data frames.
}
\examples{
data(iris)
library(magrittr)
condformat(iris,
           select=c("Sepal.Length", "Species"),
           col.names=c("Sepal length", "Species")) \%>\% print
}

