% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat_render.R
\name{condformat2excel}
\alias{condformat2excel}
\title{Writes the table to an Excel sheet}
\usage{
condformat2excel(x, filename, ...)
}
\arguments{
\item{x}{A condformat_tbl object}

\item{filename}{The xlsx file name}
}
\description{
Writes the table to an Excel sheet
}

