% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_gtable.R
\name{condformat2grob}
\alias{condformat2grob}
\title{Converts the table to a grid object}
\usage{
condformat2grob(x, draw = TRUE)
}
\arguments{
\item{x}{A condformat_tbl object}

\item{draw}{Boolean value (default TRUE). If TRUE, the table is
immediately drawn using grid::draw and the grob is returned.
If FALSE, the grob is returned without drawing. Set draw=FALSE
when using the grob in composite images with grid.arrange/ggarrange.}
}
\value{
the grid object
}
\description{
Converts the table to a grid object
}
\examples{
library(condformat)
data.frame(Student = c("Alice", "Bob", "Charlie"),
           Evaluation = c("Great", "Well done", "Good job!")) \%>\%
 condformat \%>\%
 condformat2grob
}
