% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat_render.R
\name{condformat2latex}
\alias{condformat2latex}
\title{Converts the table to LaTeX code}
\usage{
condformat2latex(x, ...)
}
\arguments{
\item{x}{A condformat_tbl object}

\item{...}{arguments passed to knitr::kable}
}
\value{
A character vector of the table source code
}
\description{
Converts the table to LaTeX code
}

