% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_latex.R
\name{condformat_css_tolatex}
\alias{condformat_css_tolatex}
\title{How to export css values to latex}
\usage{
condformat_css_tolatex(css_values)
}
\arguments{
\item{css_values}{A character matrix with the CSS values that need to
be converted to Latex}
}
\value{
A list with two character matrices named `before` and `after`. Both
        of these matrices must be of the same size as `css_values`.
}
\description{
How to export css values to latex
}
\examples{
\dontrun{
# This code implements the piece needed for converting the font-weight CSS
# property. The before matrix contains `"\\textbf leftbrace"` in the positions where a bold
# text is expected. In those cases the "after" matrix contains `"rightbrace"`.
# Therefore the final value will be paste0("\\\\textbf leftbrace", value, "rightbrace")
`condformat_css_tolatex.font-weight` <- function(css_values) {
  # \\textbf{setosa}
  before <- ifelse(css_values == "bold", "\\textbf{", "")
  after <- ifelse(css_values == "bold", "}", "")
  list(before=before, after=after)
}
}
}
