% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule_helper.R
\name{get_css_field}
\alias{get_css_field}
\title{Returns a matrix with the value of the CSS field for each element
of our data frame.}
\usage{
get_css_field(finalformat, field)
}
\arguments{
\item{finalformat}{list with the css_fields}

\item{field}{the name of the CSS field to be returned}
}
\value{
a matrix of size xview with the CSS field information
}
\description{
Returns a matrix with the value of the CSS field for each element
of our data frame.
}

