% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat_render.R
\name{knit_print.condformat_tbl}
\alias{knit_print.condformat_tbl}
\title{Print method for knitr, exporting to HTML or LaTeX as needed}
\usage{
\method{knit_print}{condformat_tbl}(x, ...)
}
\arguments{
\item{x}{Object to print}

\item{...}{On a LaTeX output these are unused. On an HTML output can have \code{paginate=TRUE} or \code{paginate = FALSE}}
}
\description{
Print method for knitr, exporting to HTML or LaTeX as needed
}
