% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condformat.R
\name{+.condformat_tbl}
\alias{+.condformat_tbl}
\title{Prints the data frame in an html page and shows it.}
\usage{
\method{+}{condformat_tbl}(x, rule)
}
\arguments{
\item{x}{A condformat_tbl object}

\item{rule}{The rule to be applied}
}
\value{
x, with an attribute defining the rule
}
\description{
Prints the data frame in an html page and shows it.
}
\examples{
data(iris)
condformat(iris) + rule_column("Species", "background:red")
}

