% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{uq}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lazyeval}{\code{\link[lazyeval]{uq}}}
}}
\keyword{internal}

