% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule_column.R
\name{rule_column}
\alias{rule_column}
\title{Formats a whole column of a condformat_tbl}
\usage{
rule_column(column, css, lockcells = FALSE)
}
\arguments{
\item{column}{a character string or vector with the column names to be coloured.}

\item{css}{a character string with the CSS code to be applied to those columns}

\item{lockcells}{logical value determining if no further rules should be applied to the affected cells.}

\item{x}{A condformat_tbl, or an object that can be coerced to that}
}
\value{
The condformat_tbl object, with the added information
}
\description{
Formats a whole column of a condformat_tbl
}
\examples{
data(iris)
library(magrittr)
condformat(iris) \%>\% rule_column(column="Species", css="background: red")
}

