% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule_highlight.R
\name{rule_highlight}
\alias{rule_highlight}
\title{Highlights cells of a column based on a given condition}
\usage{
rule_highlight(column, condition, css, lockcells = FALSE)
}
\arguments{
\item{column}{a character string or vector with the column names to be coloured.}

\item{condition}{a character string with the condition to be evaluated to apply the given css format.}

\item{css}{a character string with the CSS code to be applied to those columns}

\item{lockcells}{logical value determining if no further rules should be applied to the affected cells.}
}
\value{
The condformat_tbl object, with the added information
}
\description{
Highlights cells of a column based on a given condition
}
\examples{
data(iris)
library(magrittr)
condformat(iris) \%>\%
  rule_highlight(column="Species",
                 condition="Sepal.Width > Sepal.Length - 2.25",
                 css="background: #7D00FF") \%>\% print
}

