% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_columns.R
\name{show_columns_}
\alias{show_columns_}
\title{Show columns (deprecated)}
\usage{
show_columns_(..., .dots, col_names)
}
\arguments{
\item{...}{Comma separated list of unquoted expressions}

\item{.dots}{A character vector with columns to show}

\item{col_names}{Character vector with the column names for the selected columns}
}
\description{
Show columns (deprecated)
}
\examples{

data(iris)
x <- head(iris)
# Use standard evaluation (columns as strings):
condformat(x) +
  show_columns_(.dots = c("Sepal.Length", "Species"), col_names = c("Sepal Length", "Species"))
}
