% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rows.R
\name{show_rows_old}
\alias{show_rows_old}
\alias{show_rows_}
\title{Selects the rows to be printed (deprecated)}
\usage{
show_rows_old(...)

show_rows_(..., .dots)
}
\arguments{
\item{...}{Expressions used for filtering}

\item{.dots}{A list of lazy objects. See examples}
}
\value{
A condformat_show_rows object, usually to be added to a condformat_tbl object
as shown in the examples
}
\description{
This function is deprecated. Use \code{\link{show_rows}} instead
}
\details{
Keeps the rows you mention in the printed table.
Compared to \code{\link[dplyr]{filter}}, show_rows does not remove the
rows from the actual data frame, they are removed only for printing.
}
\examples{
library(condformat)
data(iris)
x <- head(iris)
condformat(x) + show_rows(Sepal.Length > 4.5, Species == "setosa")
library(condformat)
data(iris)
x <- head(iris)
condformat(x) + show_rows_(.dots = c("Sepal.Length > 4.5", "Species == 'setosa'"))
}
\seealso{
\code{\link[dplyr]{filter}}
}
