% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_caption.R
\name{theme_caption}
\alias{theme_caption}
\title{Sets the caption of a condformat object}
\usage{
theme_caption(x, caption = "")
}
\arguments{
\item{x}{The condformat object}

\item{caption}{The caption to show}
}
\description{
The advantage with respect to theme_htmlTable(caption = "My table") is that
this works with HTML and LaTeX outputs
}
\examples{
data(iris)
condformat(head(iris)) \%>\%
  theme_caption(caption = "My Caption")
}
