% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grob.R
\name{theme_grob}
\alias{theme_grob}
\title{Customizes appearance of condformat object}
\usage{
theme_grob(x, ...)
}
\arguments{
\item{x}{The condformat object}

\item{...}{Arguments to be passed to gridExtra::tableGrob (see examples)}
}
\description{
This is only used on grob output.
}
\examples{
data(iris)
cf <- condformat(head(iris)) \%>\%
  theme_grob(
    rows = NULL,
    theme = gridExtra::ttheme_default(base_size = 10, base_colour = "red")
  )
condformat2grob(cf)
}
\seealso{
\code{\link[gridExtra]{tableGrob}}
}
