% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_htmlWidget.R
\name{theme_htmlWidget}
\alias{theme_htmlWidget}
\title{Customizes appearance of condformat object}
\usage{
theme_htmlWidget(x, ...)
}
\arguments{
\item{x}{The condformat object}

\item{...}{Arguments to be passed to htmlTable::htmlTableWidget (see examples)}
}
\description{
Customizes appearance of condformat object
}
\examples{
data(iris)
condformat(head(iris)) \%>\%
  theme_htmlWidget(number_of_entries = c(10, 25, 100),
                   width = NULL, height = NULL, elementId = NULL)
}
\seealso{
\code{\link[htmlTable]{htmlTable}}
}
