% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kable.R
\name{theme_kable}
\alias{theme_kable}
\title{Customizes appearance of condformat object}
\usage{
theme_kable(x, ...)
}
\arguments{
\item{x}{The condformat object}

\item{...}{Arguments to be passed to knitr::kable (see examples)}
}
\description{
This is only used on LaTeX output.
}
\examples{
data(iris)
cf <- condformat(head(iris)) \%>\%
  theme_kable(booktabs = TRUE, caption = "My Caption")
\dontrun{
print(cf)
}
}
\seealso{
\code{\link[knitr]{kable}}
}
