% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{condformat-shiny}
\alias{condformat-shiny}
\alias{condformatOutput}
\alias{renderCondformat}
\title{Shiny bindings for condformat}
\usage{
condformatOutput(outputId, ...)

renderCondformat(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{...}{arguments passed to htmlOutput}

\item{expr}{An expression that generates a condformat object}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using condformat within Shiny
applications and interactive Rmd documents.
}
\details{
Pagination will be used if a recent enough htmlTable package is available.
}

