% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat.R
\name{condformat}
\alias{condformat}
\title{Converts a data frame or matrix to a condformat_tbl class.}
\usage{
condformat(x)
}
\arguments{
\item{x}{A matrix or data.frame}
}
\value{
The condformat_tbl object. This object can be used
        in a ggplot-like syntax to apply conditional formatting
        rules. It can also be used as a conventional
        data frame.

        The condformat_tbl print method generates an htmlTable, to be
        viewed using RStudio Viewer or an HTML browser, as available.
}
\description{
condformat_tbl objects allow to add conditional formatting information,
that can be viewed when the condformat_tbl object is printed.
}
\details{
They also have all properties from conventional data frames.
}
\examples{
data(iris)
condformat(iris[1:5,])
}

