% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_gtable.R
\name{condformat2grob}
\alias{condformat2grob}
\title{Converts the table to a grid object}
\usage{
condformat2grob(x)
}
\arguments{
\item{x}{A condformat_tbl object}
}
\value{
the grid object
}
\description{
Converts the table to a grid object
}
\examples{
library(condformat)
data.frame(Student = c("Alice", "Bob", "Charlie"),
           Evaluation = c("Great", "Well done", "Good job!")) \%>\%
 condformat \%>\%
 condformat2grob
}
