% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat.R
\name{+.condformat_tbl}
\alias{+.condformat_tbl}
\title{Combines the data with the formatting rules and graphical options.}
\usage{
\method{+}{condformat_tbl}(x, obj)
}
\arguments{
\item{x}{A condformat_tbl object}

\item{obj}{A condformat_show or a condformat_rule object to be combined
Any other type of object will be added as expected to the data frame.}
}
\value{
x, with extended condformat_tbl attributes
}
\description{
Combines the data with the formatting rules and graphical options.
}
\examples{
data(iris)
condformat(iris) + show_columns(Species)
}

