% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat_render.R
\name{print.condformat_tbl}
\alias{print.condformat_tbl}
\title{Prints the data frame in an html page and shows it.}
\usage{
\method{print}{condformat_tbl}(x, ..., paginate = TRUE)
}
\arguments{
\item{x}{A condformat_tbl object}

\item{...}{optional arguments to print}
}
\value{
the value returned by htmlTable
}
\description{
Prints the data frame in an html page and shows it.
}
\examples{
data(iris)
print(condformat(iris[1:5,]))
}
