% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condformat_render.R
\name{print.condformat_tbl}
\alias{print.condformat_tbl}
\title{Prints the data frame in an html page and shows it.}
\usage{
\method{print}{condformat_tbl}(x, ..., paginate = TRUE)
}
\arguments{
\item{x}{A condformat_tbl object}

\item{...}{Arguments passed on to \code{htmltools::html_print}
\describe{
  \item{background}{Background color for web page}
  \item{viewer}{A function to be called with the URL or path to the generated
HTML page. Can be \code{NULL}, in which case no viewer will be invoked.}
}}

\item{paginate}{A logical value. If TRUE the printing will be paginated}
}
\value{
the value returned by htmlTable
}
\description{
Prints the data frame in an html page and shows it.
}
\examples{
data(iris)
print(condformat(iris[1:5,]))
}
