% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule_fill_gradient.R
\name{rule_fill_gradient}
\alias{rule_fill_gradient}
\title{Fills the background color of a column using a gradient based on
the values of a given column}
\usage{
rule_fill_gradient(x, column, colour_by = column, low = "#132B43",
  high = "#56B1F7", space = "Lab", na.value = "#7F7F7F",
  limits = range(x[, colour_by], na.rm = TRUE), lockcells = FALSE)
}
\arguments{
\item{x}{A condformat_tbl object.}

\item{column}{a character string or vector with the column names to be coloured.}

\item{colour_by}{a character string with the column name of the values to be used to generate the gradient}

\item{low,high}{a character string with the CSS color code}

\item{space}{the color space passed to scales::seq_gradient_pal}

\item{na.value}{a character string with the CSS color to be used in missing values}

\item{limits}{range of limits that the gradient should cover}

\item{lockcells}{logical value determining if no further rules should be applied to the affected cells.}
}
\value{
The condformat_tbl object, with the added gradient information
}
\description{
Fills the background color of a column using a gradient based on
the values of a given column
}
\examples{
data(iris)
library(magrittr)
condformat(iris) \%>\% rule_fill_gradient(column="Sepal.Length") \%>\% print
}

