% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule_fill_gradient2.R
\name{rule_fill_gradient2}
\alias{rule_fill_gradient2}
\title{Fills the background color of a column using a three colors gradient based on
the values of a given column}
\usage{
rule_fill_gradient2(column, colour_by = column, low = scales::muted("red"),
  mid = "white", high = scales::muted("blue"), midpoint = NA,
  space = "rgb", na.value = "#7F7F7F", limits = NA, lockcells = FALSE)
}
\arguments{
\item{column}{a character string or vector with the column names to be coloured.}

\item{colour_by}{a character string with the column name of the values to be used to generate the gradient}

\item{low,mid,high}{a character string with the CSS color code for the three levels of the gradient}

\item{midpoint}{the value used for the middle color}

\item{space}{the color space passed to scales::div_gradient_pal}

\item{na.value}{a character string with the CSS color to be used in missing values}

\item{limits}{range of limits that the gradient should cover}

\item{lockcells}{logical value determining if no further rules should be applied to the affected cells.}
}
\value{
The condformat_tbl object, with the added gradient information
}
\description{
Fills the background color of a column using a three colors gradient based on
the values of a given column
}
\examples{
data(iris)
library(magrittr)
condformat(iris) \%>\% rule_fill_gradient(column="Sepal.Length") \%>\% print
}

