% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_columns.R
\name{show_columns}
\alias{show_columns}
\title{Selects the variables to be printed}
\usage{
show_columns(..., col_names)
}
\arguments{
\item{...}{Comma separated list of unquoted expressions}

\item{col_names}{Character vector with the column names for the selected columns}
}
\value{
A condformat_show_columns object, usually to be added to a condformat_tbl object
}
\description{
Keeps the variables you mention in the printed table.
Compared to \code{\link[dplyr]{select}}, show_columns does not remove the
columns from the data frame, so formatting rules can still depend
on them.
}
\examples{

library(dplyr) # for starts_with()
data(iris)
x <- head(iris)

# Include some columns:
condformat(x) + show_columns(Sepal.Length, Sepal.Width, Species)

# Rename columns:
condformat(x) + show_columns(Sepal.Length, Species, col_names = c("Length", "Spec."))

# Exclude some columns:
condformat(x) + show_columns(-Petal.Length, -Petal.Width)

# Select columns using dplyr syntax:
condformat(x) + show_columns(starts_with("Petal"), Species)

}
\seealso{
\code{\link[dplyr]{select}}
}
