% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_htmlTable.R
\name{theme_htmlTable}
\alias{theme_htmlTable}
\title{Customizes appearance of condformat object}
\usage{
theme_htmlTable(x, ...)
}
\arguments{
\item{x}{The condformat object}

\item{...}{Arguments to be passed to htmlTable}
}
\description{
Customizes appearance of condformat object
}
\examples{
data(iris)
condformat(head(iris)) \%>\% theme_htmlTable(caption="Table 1: My iris table", rnames=FALSE)
}
\seealso{
\code{\link[htmlTable]{htmlTable}}
}
